+++ Indicators/WilliamR/WilliamR.Models.cs
@@ -3,7 +3,7 @@
 
     public class WilliamResult : ResultBase
     {
-        public float? WilliamR { get; set; }
+        public decimal? WilliamR { get; set; }
         public bool? IsIncreasing { get; set; }
     }
